Second attempt at the large commit to add in the new hilltop flow routing functions. The run time has been reduced from ~40 hrs to ~2 hrs on my test dataset over the old code and we are now getting a much higher percentage of hilltop pixels sucessfully being routed.
The hfr code is called using the hfr_driver which calls a couple of 
recursive functions to cycle through all the pixels in the landcscape. See 
the documentation for details on the output data.

Also added a new function to statstools which is needed by the HFR which converts a
bearing in degrees into radians, rotating 0 from north to east.